NIF: return updated length when adding an element to a batch
This solves the issue presented in the second comment of #8.
Now if two processes simultaneously try to add an Account/Transfer calling the
Elixir API, each of them will get a different new_length, and they will set the
fields on their own new_length - 1 respectively, without stomping on each others
feet.
Issue #8 still remains open because we can still evaluate appending binaries to
batches in a single shot, but this gives us more time to consider that since now
the code is at least correct.