docker-compose: build the Docker image locally
Since we already clone the TigerBeetle repo, we can build the Docker
image locally. This ensures that the runner builds it with the CPU
features it supports, avoiding failing CI due to the default CPU
baseline.
Moreover, this removes the need to manually bump TigerBeetle's version
when updating the submodule.
Fix #16.