{account,transfer}_batch: add append/2 function
Appends an item to the batch by directly serializing it on the Elixir side. The
binary is then passed to the NIF which just copies it into the newly added batch
slot.
Also add a benchmark to compare this new append/2 function with the old
add_{account,transfer}/2, before removing them.