feat!: bye bye, batches
Remove the input batch concept (i.e. a NIF resource that preallocated a binary
with a fixed size to accomodate a certain number of Events) and replace it with
a more idiomatic API.

I'll probably write a blogpost on this, but in short:
- I was lured into this trying to replicate the Java client mechanism described
in "Writing High-Performance Clients for Tigerbeetle"
- ...but the API is excessively clunky when used in a language with immutable
data
- ...and it doesn't actually make things faster in the Elixir case
- ...and it adds lots of code both on the Elixir and the Zig side

So, just use an idiomatic API for Elixir, passing lists of structs as input.
I've run the benchmark before and after and the throughput is fundamentally the
same (when measuring the same thing, the old benchmark actually "cheated" a
little).