refactor: don't hide pointer in beam.Env declaration
NASA's Power of Ten, Rule 9
[...] Pointer dereference operations may not be hidden in macro
definitions or inside typedef declarations.

By explicitly passing beam.Env as a pointer, we also are able to assert
non-nullability once at the source (when receiving, allocating the env, or
casting it from an int) and then pass around a pointer that is guaranteed to be
non-null.

A similar change for beam.ResourceType will be done when refactoring code around
NIF resources.