fix(client): allocate a NIF env for each request
Having a single shared env doesn't work if multiple callers use the same client
with the processless API, since the env gets cleared at the end of each request.
Instead, we now allocate a fresh env for each request, save it in the request
context and free it at the end. This also simplifies resource deinitialization
since the completion context is not used anymore.
Add the smoke test that was used to detect the bug to the testsuite.