fix: make TigerBeetlex work on Mac Silicon
Mac CI was green, but it was an Intel runner. Trying out TigerBeetle on
a Mac M3 highlighted a comptime assert failure in the hashing code, due
AES being disabled.

Reusing TigerBeetle's resolve_target function in build.zig fixes that.

Add back macos-14 and macos-15 runners to the CI matrix.